Implemented the "FizzBuzz" case.
I had postpone implementation of the "FizzBuzz" case for later just to
feel if adding a new case felt good in a "clean code" version of the
kata.

It is obvious that FizzBuzzGameBoard->build method will get dirtier as
new "cases" could appear in the game.

That method should be my next concern.